Rename and move CopyJsonLoggerInterface and implementation
