+++ src/ElasticSearch/JsonDocument/Properties/AddressTransformer.php
@@ -3,12 +3,12 @@
 namespace CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\Properties;
 
 use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\CopyJsonInterface;
-use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\Logging\CopyJsonLoggerInterface;
+use CultuurNet\UDB3\Search\JsonDocument\JsonTransformerLogger;
 
 class AddressTransformer implements CopyJsonInterface
 {
     /**
-     * @var CopyJsonLoggerInterface
+     * @var JsonTransformerLogger
      */
     private $logger;
 
@@ -17,7 +17,7 @@ class AddressTransformer implements CopyJsonInterface
      */
     private $addressRequired;
 
-    public function __construct(CopyJsonLoggerInterface $logger, $addressRequired)
+    public function __construct(JsonTransformerLogger $logger, $addressRequired)
     {
         $this->logger = $logger;
         $this->addressRequired = $addressRequired;
