+++ src/ElasticSearch/JsonDocument/Properties/AvailabilityTransformer.php
@@ -5,18 +5,18 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\Properties;
 
 use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\CopyJsonInterface;
-use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\Logging\CopyJsonLoggerInterface;
+use CultuurNet\UDB3\Search\JsonDocument\JsonTransformerLogger;
 use DateTimeImmutable;
 use stdClass;
 
 final class AvailabilityTransformer implements CopyJsonInterface
 {
     /**
-     * @var CopyJsonLoggerInterface
+     * @var JsonTransformerLogger
      */
     private $copyLogger;
 
-    public function __construct(CopyJsonLoggerInterface $copyLogger)
+    public function __construct(JsonTransformerLogger $copyLogger)
     {
         $this->copyLogger = $copyLogger;
     }
