+++ src/ElasticSearch/JsonDocument/Properties/CreatorTransformer.php
@@ -3,21 +3,21 @@
 namespace CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\Properties;
 
 use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\CopyJsonInterface;
-use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\Logging\CopyJsonLoggerInterface;
+use CultuurNet\UDB3\Search\JsonDocument\JsonTransformerLogger;
 
 class CreatorTransformer implements CopyJsonInterface
 {
     /**
-     * @var CopyJsonLoggerInterface
+     * @var JsonTransformerLogger
      */
     private $logger;
 
     /**
      * CopyJsonCreator constructor.
      *
-     * @param CopyJsonLoggerInterface $logger
+     * @param JsonTransformerLogger $logger
      */
-    public function __construct(CopyJsonLoggerInterface $logger)
+    public function __construct(JsonTransformerLogger $logger)
     {
         $this->logger = $logger;
     }
