+++ src/ElasticSearch/JsonDocument/Properties/NameTransformer.php
@@ -3,20 +3,20 @@
 namespace CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\Properties;
 
 use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\CopyJsonInterface;
-use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\Logging\CopyJsonLoggerInterface;
+use CultuurNet\UDB3\Search\JsonDocument\JsonTransformerLogger;
 
 class NameTransformer implements CopyJsonInterface
 {
     /**
-     * @var CopyJsonLoggerInterface
+     * @var JsonTransformerLogger
      */
     private $logger;
 
     /**
      * CopyJsonName constructor.
-     * @param CopyJsonLoggerInterface $logger
+     * @param JsonTransformerLogger $logger
      */
-    public function __construct(CopyJsonLoggerInterface $logger)
+    public function __construct(JsonTransformerLogger $logger)
     {
         $this->logger = $logger;
     }
