+++ src/ElasticSearch/JsonDocument/Properties/WorkflowStatusTransformer.php
@@ -3,13 +3,13 @@
 namespace CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\Properties;
 
 use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\CopyJsonInterface;
-use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\CopyJson\Logging\CopyJsonLoggerInterface;
+use CultuurNet\UDB3\Search\JsonDocument\JsonTransformerLogger;
 use stdClass;
 
 class WorkflowStatusTransformer implements CopyJsonInterface
 {
     /**
-     * @var CopyJsonLoggerInterface
+     * @var JsonTransformerLogger
      */
     private $logger;
 
@@ -18,7 +18,7 @@ class WorkflowStatusTransformer implements CopyJsonInterface
      */
     private $default;
 
-    public function __construct(CopyJsonLoggerInterface $logger, ?string $default = null)
+    public function __construct(JsonTransformerLogger $logger, ?string $default = null)
     {
         $this->logger = $logger;
         $this->default = $default;
